Fixed SD2-586: Debugger not going to source code line from callstack view
git-svn-id: svn://svn.sharpdevelop.net/sharpdevelop/trunk@896 1ccf3a8d-04fe-1044-b7c0-cef0b8235c61

+++ src/AddIns/Misc/Debugger/Debugger.AddIn/Project/Src/Service/WindowsDebugger.cs
@@ -414,7 +414,9 @@ namespace ICSharpCode.SharpDevelop.Services
 		void DebuggingResumed(object sender, DebuggerEventArgs e)
 		{
 			isProcessRunningCache = true;
-			DebuggerService.RemoveCurrentLineMarker();
+			if (!debugger.Evaluating) {
+				DebuggerService.RemoveCurrentLineMarker();
+			}
 		}
 
 		public void JumpToCurrentLine()
